Langchain, Langgraph e Fast API/Assistente/backend/requirements.txt
fastapi
uvicorn
langchain
langchain_openai
openai
python-dotenv
